Rename desktop platform to desktop-sdl2
