+++ src/platforms/desktop-sdl2/platform.c
@@ -20,7 +20,7 @@
 #include "../../platform.h"
 #include "../../sounds.h"
 
-#include "../../modules/platforms/desktop.h"
+#include "../../modules/platforms/desktop_sdl2.h"
 
 #define FPS 60
 #define FRAME_TIME_LENGTH (1000 / FPS)
